+++ Models/SysDaftar.cs
@@ -5,7 +5,7 @@ namespace CityInfoApi.Models_new
 {
     public partial class SysDaftar
     {
-        public long DId { get; set; }
+        public int DId { get; set; }
         public string? DName { get; set; }
 
     }
